Add simple outputter
